0.7.48: polyfonie tot 32.768, uitgangen tot 1.024, samplerate kiezen
De drie omcirkelde getallen uit de Hauptwerk-editietabel.

Polyfonie: plafond 4.096 -> 32.768. Wel eerlijk gemeten wat een pc haalt,
want een plafond is geen belofte. Op een i9-10885H (WASAPI, 48 kHz, 480
frames, Friesach met alle 44 registers): 210 stemmen 29%, 424 stemmen 68%,
828 stemmen 139% (over de deadline), 1088 stemmen 195%. Dus ~0,17% per
stem, en rond 500 stemmen loopt het renderen tegen zijn grens. Het echte
plafond is de ene kern waarop de mengloop draait.

Uitgangen: MAX_OUT_CH 64 -> 1024, en belangrijker: de kanaalnummers waren
een u8, dus kanaal 255 was het hoogste dat een divisie kon aanwijzen. Nu
u16 van de UI tot de audiothread en het opgeslagen instellingenbestand.

Samplerate: nieuw veld AudioOutputConfig::sample_rate. Wordt alleen
overgenomen als het apparaat hem aanbiedt (supported_output_configs);
anders blijft de huidige staan met een waarschuwing in het log in plaats
van een mislukte stream. Keuzelijst in de audio-instellingen, bewaard
tussen sessies. Gemeten: 96 kHz gevraagd op de Realtek-uitgang onder
WASAPI gedeelde modus -> netjes geweigerd, 48 kHz blijft staan. Dat is een
beperking van WASAPI shared, niet van ons; de hint zegt waar je het dan
wel instelt (Windows-geluidsinstellingen of het ASIO-paneel).

Drie nieuwe tests voor de kanaalweging (incl. kanalen boven 255).

+++ VirtualPipeOrgan/src-tauri/src/library.rs
@@ -156,8 +156,10 @@ pub struct DivisionOutputPairSaved {
     pub division: String,
     #[serde(default)]
     pub pair: u8,
+    /// Uitgangskanalen van deze divisie. u16 sinds 0.7.48: met u8 was kanaal
+    /// 255 het hoogste dat een divisie kon aanwijzen.
     #[serde(default)]
-    pub channels: Option<Vec<u8>>,
+    pub channels: Option<Vec<u16>>,
 }
 
 /// C/Cis-lade spreiding aan/uit per divisie
